Move shared gui element out of controller package.
Move shared gui element from controller package to shared gui package.
