+++ src/main/java/Controller/MainController.java
@@ -1,5 +1,8 @@
 package Controller;
 
+import SharedGuiElement.OpenCode;
+import SharedGuiElement.OpenCodeBuilder;
+import SharedGuiElement.RemainingTimeController;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.util.Callback;
