+++ src/main/java/SharedGuiElement/OpenCode.java
@@ -1,4 +1,4 @@
-package Controller;
+package SharedGuiElement;
 
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
