+++ src/main/java/SharedGuiElement/OpenCodeController.java
@@ -1,4 +1,4 @@
-package Controller;
+package SharedGuiElement;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.TextArea;
