+++ src/main/java/SharedGuiElement/RemainingTimeController.java
@@ -1,4 +1,4 @@
-package Controller;
+package SharedGuiElement;
 
 import javafx.fxml.FXML;
 import javafx.scene.text.Text;
