+++ src/main/java/SharedGuiElement/SubmissionTableController.java
@@ -1,7 +1,6 @@
-package Controller;
+package SharedGuiElement;
 
 import javafx.collections.FXCollections;
-import javafx.collections.ListChangeListener;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
