+++ src/main/resources/SharedGuiElement/OpenCode.fxml
@@ -2,7 +2,7 @@
 <?import javafx.scene.control.TextArea?>
 <?import javafx.geometry.Insets?>
 <VBox xmlns:fx="http://javafx.com/fxml"
-      fx:controller="Controller.OpenCodeController" >
+      fx:controller="SharedGuiElement.OpenCodeController" >
     <padding><Insets top="25" right="25" bottom="25" left="25"/></padding>
     <TextArea fx:id="code"
               editable="false" />
