+++ src/main/resources/SharedGuiElement/QuestionAndAnswerTable.fxml
@@ -1,7 +1,7 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.HBox?>
 <HBox xmlns:fx="http://javafx.com/fxml"
-      fx:controller="Controller.QuestionAndAnswerTableController" >
+      fx:controller="SharedGuiElement.QuestionAndAnswerTableController" >
     <TableView fx:id="questionAndAnswerTable"
                HBox.hgrow="ALWAYS" >
         <columns>
