+++ src/main/resources/SharedGuiElement/RemainingTime.fxml
@@ -3,7 +3,7 @@
 <?import javafx.scene.layout.GridPane?>
 <?import javafx.scene.text.Text?>
 <GridPane xmlns:fx="http://javafx.com/fxml"
-          fx:controller="Controller.RemainingTimeController"
+          fx:controller="SharedGuiElement.RemainingTimeController"
           alignment="center"
           hgap="10"
           vgap="10" >
